updated based on feedback
