+++ packages/webawesome/scripts/build.js
@@ -16,7 +16,7 @@ import { SimulateWebAwesomeApp } from '../docs/_utils/simulate-webawesome-app.js
 import { generateDocs } from './docs.js';
 import { generateLlmsTxtFile } from './llms.js';
 import { getCdnDir, getDistDir, getDocsDir, getRootDir, getSiteDir } from './utils.js';
-import { verifyCem } from './verify-cem.js';
+
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 
@@ -60,7 +60,7 @@ export async function build(options = {}) {
     const start = Date.now();
 
     try {
-      const steps = [cleanup, generateManifest, verifyManifest, generateReactWrappers, generateTypes, generateStyles];
+      const steps = [cleanup, generateManifest, generateReactWrappers, generateTypes, generateStyles];
 
       for (const step of steps) {
         if (debugPerf) {
@@ -107,7 +107,7 @@ export async function build(options = {}) {
   /**
    * Analyzes components and generates the custom elements manifest file.
    */
-  function generateManifest() {
+  async function generateManifest() {
     spinner.start('Generating CEM');
 
     try {
@@ -120,19 +120,41 @@ export async function build(options = {}) {
       }
     }
 
-    spinner.succeed();
-
-    return Promise.resolve();
-  }
-
-  /**
-   * Verifies the custom elements manifest doesn't have issues like unnamed events.
-   */
-  async function verifyManifest() {
-    spinner.start('Verifying CEM');
-
+    // Verify the manifest doesn't have issues like unnamed events
+    // See: https://github.com/shoelace-style/webawesome/issues/1919
     try {
-      await verifyCem();
+      const errors = [];
+      const cemPath = join(getCdnDir(), 'custom-elements.json');
+      const cem = JSON.parse(await readFile(cemPath, 'utf-8'));
+
+      for (const module of cem.modules || []) {
+        for (const declaration of module.declarations || []) {
+          if (declaration.kind === 'class' && declaration.events) {
+            for (const event of declaration.events) {
+              if (!event.name) {
+                errors.push(
+                  `Component "${declaration.name}" has an event without a name (type: ${event.type?.text || 'unknown'}). ` +
+                    `This will generate "onundefined" in JSX types. Add an @event JSDoc tag with the event name.`
+                );
+              }
+            }
+          }
+        }
+      }
+
+      const jsxTypesPath = join(getCdnDir(), 'custom-elements-jsx.d.ts');
+      const jsxTypes = await readFile(jsxTypesPath, 'utf-8');
+
+      if (jsxTypes.includes('onundefined')) {
+        errors.push(
+          'custom-elements-jsx.d.ts contains "onundefined" event handlers. ' +
+            'This indicates events are missing names in the Custom Elements Manifest.'
+        );
+      }
+
+      if (errors.length > 0) {
+        throw new Error(`CEM verification failed:\n\n${errors.map(e => `  - ${e}`).join('\n')}`);
+      }
     } catch (error) {
       spinner.fail();
       console.error(`\n\n${error.message}`);
